Support undefined values for props.
Previously, as the props value was required, the (compound)variant(s)
matching the default value were applyed.

If no style was expected, a "default" variant with an empty declaration
was required.

Now, you only need to define props as optional and if no fallback is
passed, no style will be applyed.